% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{filter}
\alias{filter}
\title{filter combinator which returns filtered vector by predicate}
\usage{
filter(x, p)
}
\arguments{
\item{x}{vector to filter}

\item{p}{predicate}
}
\value{
filtered vector
}
\description{
filter combinator which returns filtered vector by predicate
}
\examples{
filter(c(1, 2, 3, 4, 5, 6), (~ x > 3))
}
