% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{filter_not}
\alias{filter_not}
\title{filter_not combinator which returns filtered vector by reversed predicate}
\usage{
filter_not(x, p)
}
\arguments{
\item{x}{vector to filter}

\item{p}{predicate}
}
\value{
filtered vector
}
\description{
filter_not combinator which returns filtered vector by reversed predicate
}
\examples{
filter_not(c(1, 2, 3, 4, 5, 6), (~ x > 3))
}
